+++ src/iqm/benchmarks/entanglement/ghz.py
@@ -19,6 +19,7 @@ GHZ state benchmark
 from itertools import chain
 from time import strftime
 from typing import Any, Dict, List, Optional, Sequence, Tuple, Type, cast
+import warnings
 
 from matplotlib.figure import Figure
 import matplotlib.pyplot as plt
@@ -27,7 +28,6 @@ from networkx import Graph, all_pairs_shortest_path, is_connected, minimum_spann
 import numpy as np
 from qiskit import ClassicalRegister, QuantumRegister
 from qiskit.quantum_info import random_clifford
-from qiskit.transpiler import CouplingMap
 from qiskit_aer import Aer
 from scipy.spatial.distance import hamming
 import xarray as xr
@@ -45,7 +45,7 @@ from iqm.benchmarks.circuit_containers import BenchmarkCircuit, CircuitGroup, Ci
 from iqm.benchmarks.logging_config import qcvv_logger
 from iqm.benchmarks.readout_mitigation import apply_readout_error_mitigation
 from iqm.benchmarks.utils import (
-    extract_fidelities,
+    extract_fidelities_unified,
     perform_backend_transpilation,
     reduce_to_active_qubits,
     retrieve_all_counts,
@@ -309,7 +309,7 @@ def generate_ghz_log_cruz(num_qubits: int) -> QuantumCircuit:
 
 
 def generate_ghz_star_optimal(
-    qubit_layout: List[int], cal_url: str, backend: IQMBackendBase, inv: bool = False
+    qubit_layout: List[int], iqm_server_url: str, backend: IQMBackendBase, quantum_computer: str, inv: bool = False,
 ) -> QuantumCircuit:
     """
     Generates the circuit for creating a GHZ state by maximizing the number of CZ gates between a pair of MOVE gates.
@@ -317,10 +317,12 @@ def generate_ghz_star_optimal(
     Args:
         qubit_layout: List[int]
             The layout of qubits for the GHZ state.
-        cal_url: str
-            The calibration URL for extracting fidelities.
+        iqm_server_url: str
+            The IQM server URL for extracting fidelities.
         backend: IQMBackendBase
             The backend to be used for the quantum circuit.
+        quantum_computer: str
+            The name of the quantum computer to be used for calbration data, i.e. "garnet", "emerald", ...
         inv: bool
             Whether to generate the inverse circuit.
 
@@ -330,48 +332,47 @@ def generate_ghz_star_optimal(
     num_qubits = len(qubit_layout)
 
     # Initialize quantum and classical registers
-    comp_r = QuantumRegister(1, "comp_r")  # Computational resonator
     q = QuantumRegister(backend.num_qubits, "q")  # Qubits
     c = ClassicalRegister(num_qubits, "c")
-    qc = QuantumCircuit(comp_r, q, c, name="GHZ_star_optimal")
-
-    cal_data = extract_fidelities(cal_url, all_metrics=True)
+    qc = QuantumCircuit(q, c, name="GHZ_star_optimal")
+    # Extract calibration data
+    cal_data = extract_fidelities_unified(iqm_server_url, backend, quantum_computer)
     # Determine the best move qubit
-    move_dict = {q + 1: cal_data[1][q] for q in qubit_layout}  ## +1 to match qubit indexing in cal data
+    double_move_fidelities = {k[0]: v for k, v in list(cal_data[-1]["double_move_gate_fidelity"].items())[::2]}
+    move_dict = {q: double_move_fidelities[q + 1] for q in qubit_layout}  ## +1 to match qubit indexing in cal data
+    # move_dict = {q + 1: cal_data[1][q] for q in qubit_layout}  ## +1 to match qubit indexing in cal data
     best_move = max(move_dict, key=move_dict.get)
 
-    T2 = cal_data[-1]["t2_echo_time"]
-    t2_dict = {qubit + 1: T2[qubit + 1] for qubit in qubit_layout}  ## +1 to match qubit indexing in cal data
+    T2 = cal_data[-1]["t2_time"]
+    t2_dict = {qubit: T2[qubit + 1] for qubit in qubit_layout}  ## +1 to match qubit indexing in cal data
     cz_order = dict(sorted(t2_dict.items(), key=lambda item: item[1], reverse=True))
     qubits_to_measure = list(cz_order.keys())
     cz_order.pop(best_move)
 
     # Construct the quantum circuit
-    qc.h(best_move)
-    qc.move(best_move, 0)
+    qc.r(np.pi/2, 3*np.pi/2, best_move)
+    qc.barrier(best_move)
     for qubit in cz_order.keys():
-        qc.cx(0, qubit)
-        qc.barrier()
-    qc.move(best_move, 0)
+        qc.barrier(qubit)
+        qc.r(np.pi/2, 3*np.pi/2, qubit)
+        qc.cz(best_move, qubit)
+        qc.r(np.pi/2, 5*np.pi/2, qubit)
+        qc.barrier(qubit)
     qc.barrier()
     qc.measure(sorted(qubits_to_measure), list(range(num_qubits)))
-
+    qcvv_logger.info(f"Best MOVE qubit selected: {best_move} with double MOVE fidelity {move_dict[best_move]:.4f} and T2 time {t2_dict[best_move]:.2f} us.")
     if inv:
-        comp_r = QuantumRegister(1, "comp_r")  # Computational resonator
         q = QuantumRegister(backend.num_qubits, "q")  # Qubits
         c = ClassicalRegister(num_qubits, "c")
-        qc = QuantumCircuit(comp_r, q, c, name="GHZ_star_optimal_inv")
-
-        qc.move(best_move, 0)
+        qc = QuantumCircuit(q, c, name="GHZ_star_optimal_inv")
         for qubit in reversed(cz_order.keys()):
-            qc.cx(0, qubit)
-            qc.barrier()
-        qc.move(best_move, 0)
-        qc.h(best_move)
-        qc.barrier()
-
+            qc.barrier(qubit)
+            qc.r(np.pi/2, 5*np.pi/2, qubit)
+            qc.cz(best_move, qubit)
+            qc.r(np.pi/2, 3*np.pi/2, qubit)
+            qc.barrier(qubit)
+        qc.r(np.pi/2, 3*np.pi/2, best_move)
     return qc
-
 
 def generate_ghz_star(num_qubits: int) -> QuantumCircuit:
     """
@@ -456,7 +457,7 @@ def generate_ghz_spanning_tree(
 
 
 def get_edges(
-    coupling_map: CouplingMap,
+    coupling_map: list[tuple[int]],
     qubit_layout: List[int],
     edges_cal: Optional[List[List[int]]] = None,
     fidelities_cal: Optional[List[float]] = None,
@@ -482,7 +483,7 @@ def get_edges(
     for idx, edge in enumerate(coupling_map):
         if edge[0] in qubit_layout and edge[1] in qubit_layout:
             if not set(edge) in edges_patch:
-                edges_patch.append(set(edge))
+                edges_patch.append(edge)
 
     if fidelities_cal is not None and edges_cal is not None:
         fidelities_cal = list(
@@ -491,9 +492,15 @@ def get_edges(
         fidelities_patch = []
         for edge in edges_patch:
             for idx, edge_2 in enumerate(edges_cal):
-                if edge == set(edge_2):
+                if set(edge) == set(edge_2):
                     fidelities_patch.append(fidelities_cal[idx])
-        weights = -np.log(np.array(fidelities_patch))
+        if len(fidelities_patch) != len(edges_patch):
+            warnings.warn(
+                f"Not all calibration fidelities were found for the selected qubit layout {qubit_layout}, using unweighted graph for circuit creation."
+            )
+            weights = np.ones(len(edges_patch))
+        else:
+            weights = -np.log(np.array(fidelities_patch))
     else:
         weights = np.ones(len(edges_patch))
     graph = Graph()
@@ -646,7 +653,8 @@ class GHZBenchmark(Benchmark):
         self.num_RMs = configuration.num_RMs
         self.rem = configuration.rem
         self.mit_shots = configuration.mit_shots
-        self.cal_url = configuration.cal_url
+        self.iqm_server_url = configuration.iqm_server_url
+        self.quantum_computer = configuration.quantum_computer
         self.timestamp = strftime("%Y%m%d-%H%M%S")
         self.execution_timestamp = ""
 
@@ -689,8 +697,13 @@ class GHZBenchmark(Benchmark):
                 qcvv_logger.warning(
                     f"The current backend is a star architecture for which a suboptimal state generation routine is chosen. Consider setting state_generation_routine={routine}."
                 )
-            if self.cal_url:
-                edges_cal, fidelities_cal, _, _ = extract_fidelities(self.cal_url)
+            if self.iqm_server_url is not None:
+                edges_cal, fidelities_cal, _, _, _ = extract_fidelities_unified(self.iqm_server_url, self.backend, self.quantum_computer)
+                # Replace fidelities >= 1.0 with median of fidelities < 1.0
+                valid_fidelities = [f for f in fidelities_cal if f < 1.0]
+                if valid_fidelities:
+                    median_fidelity = np.median(valid_fidelities)
+                    fidelities_cal = [f if f < 1.0 else median_fidelity for f in fidelities_cal]
                 graph = get_edges(self.backend.coupling_map, qubit_layout, edges_cal, fidelities_cal)
             else:
                 graph = get_edges(self.backend.coupling_map, qubit_layout)
@@ -718,15 +731,14 @@ class GHZBenchmark(Benchmark):
             )
             final_ghz = ghz_native_transpiled
         elif routine == "star_optimal":
-            if self.cal_url is None:
-                raise ValueError("Calibration URL must be provided for 'star_optimal' routine.")
-            ghz = generate_ghz_star_optimal(qubit_layout, self.cal_url, self.backend)
+            if self.iqm_server_url is None or self.quantum_computer is None:
+                raise ValueError("IQM server url and quantum_computer argument must be provided "
+                                 "for 'star_optimal' routine.")
+            ghz = generate_ghz_star_optimal(qubit_layout, self.iqm_server_url, self.backend, self.quantum_computer)
             circuit_group.add_circuit(ghz)
             ghz_native_transpiled = transpile_to_IQM(
                 ghz,
                 self.backend,
-                existing_moves_handling=True,
-                perform_move_routing=False,
                 optimize_single_qubits=self.optimize_sqg,
                 optimization_level=self.qiskit_optim_level,
             )
@@ -775,7 +787,7 @@ class GHZBenchmark(Benchmark):
         qc.remove_final_measurements()
         phases = [np.pi * i / (qubit_count + 1) for i in range(2 * qubit_count + 2)]
         if self.state_generation_routine == "star_optimal":
-            qc_inv = generate_ghz_star_optimal(qubit_layout, self.cal_url, self.backend, inv=True)
+            qc_inv = generate_ghz_star_optimal(qubit_layout, self.iqm_server_url, self.backend, self.quantum_computer, inv=True)
             for phase in phases:
                 qc_phase = qc.copy()
                 qc_phase.barrier()
@@ -804,8 +816,6 @@ class GHZBenchmark(Benchmark):
                 transpile_to_IQM(
                     ghz,
                     self.backend,
-                    existing_moves_handling=True,
-                    perform_move_routing=False,
                     optimize_single_qubits=self.optimize_sqg,
                     optimization_level=self.qiskit_optim_level,
                 )
@@ -1026,7 +1036,7 @@ class GHZConfiguration(BenchmarkConfigurationBase):
             * Default: True
         mit_shots (int): Total number of shots for readout error mitigation
             * Default: 1000
-        cal_url (Optional[str]): Optional URL where the calibration data for the selected backend can be retrieved from
+        iqm_server_url (Optional[str]): Optional iqm server URL where the calibration data for the selected backend can be retrieved from
             The calibration data is used for the "tree" state generation routine to prioritize couplings with high
             CZ fidelity.
             * Default: None
@@ -1043,4 +1053,4 @@ class GHZConfiguration(BenchmarkConfigurationBase):
     num_RMs: Optional[int] = 100
     rem: bool = True
     mit_shots: int = 1_000
-    cal_url: Optional[str] = None
+    iqm_server_url: Optional[str] = None