+++ src/iqm/benchmarks/utils.py
@@ -16,11 +16,13 @@
 General utility functions
 """
 from collections import defaultdict
+from enum import Enum
 from functools import wraps
 import itertools
 from math import floor
 import os
 import random
+import re
 from time import time
 from typing import Any, Dict, Iterable, List, Literal, Optional, Sequence, Set, Tuple, Union, cast
 import warnings
@@ -38,6 +40,7 @@ import requests
 import xarray as xr
 
 from iqm.benchmarks.logging_config import qcvv_logger
+from iqm.iqm_client import IQMClient
 from iqm.iqm_client.models import CircuitCompilationOptions
 from iqm.qiskit_iqm import IQMCircuit as QuantumCircuit
 from iqm.qiskit_iqm import IQMFakeDeneb, optimize_single_qubit_gates, transpile_to_IQM
@@ -286,11 +289,23 @@ def get_active_qubits(qc: QuantumCircuit) -> List[int]:
     return list(active_qubits)
 
 
-def extract_fidelities(cal_url: str, all_metrics: bool = False) -> Union[
-    Tuple[List[List[int]], List[float], str, Dict[int, int]],
-    Tuple[List[List[int]], List[float], str, Dict[int, int], Dict[str, Dict[Union[int, Tuple[int, int]], float]]],
-]:
-    """Returns couplings and CZ-fidelities from calibration data URL
+class ObservationType(Enum):
+    """Enumeration representing relevant keys to fetch for each operation in the observations."""
+
+    CZ = "cz"
+    CLIFFORD = ("cz", "clifford")
+    SQG = "prx"
+    READOUT = ("measure_fidelity", ".fidelity")
+    READOUT_QNDNESS = ("measure", "qndness", ".fidelity")
+    DOUBLE_MOVE = "move"
+    T1 = ("t1", "QB")
+    T2 = ("t2", "QB")
+
+
+def extract_fidelities_unified(
+    iqm_server_url: str, backend: IQMBackendBase, quantum_computer: str
+) -> tuple[list[list[int]], list[float], str, dict[Any, int], dict[str, dict[int | tuple[int, int], float]]]:
+    """Returns couplings and CZ-fidelities from calibration data URL for external station API
 
     Args:
         cal_url: str
@@ -304,73 +319,93 @@ def extract_fidelities(cal_url: str, all_metrics: bool = False) -> Union[
             data contains a fidelity.
         list_fids: List[float]
             A list of CZ fidelities from the calibration url, ordered in the same way as list_couplings
-        topology: str
-            Name of the chip topology layout, currently either "star" or "crystal"
-        qubit_mapping: Dict[int, int]
-            Enumerating all calibrated qubits starting from 0. For instance if on a 5 qubit chip the qubits 2, 3 are calibrated,
-            the mapping will be {2: 0, 3: 1}.
         metrics_dict: Dict
             Dictionary of all metrics (returned only if all_metrics=True)
             Format: {metric_name: {qubit: value}} for single qubit metrics
             Format: {metric_name: {(qubit_1, qubit_2): value}} for two qubit metrics
     """
-    headers = {"Accept": "application/json", "Authorization": "Bearer " + os.environ["IQM_TOKEN"]}
-    r = requests.get(cal_url, headers=headers, timeout=60)
-    calibration = r.json()
-    cal_keys = {
-        el2["key"]: (i, j) for i, el1 in enumerate(calibration["calibrations"]) for j, el2 in enumerate(el1["metrics"])
-    }
-    resonator_names = ["COMPR", "COMP_R", "MPR1", "MPR_1"]
+    # Create dictionaries to map key names to their corresponding metrics
+    cz_fidelity: Dict[Union[int, Tuple[int, int]], float] = {}
+    single_qubit_fidelity: Dict[int, float] = {}
+    readout_fidelity: Dict[int, float] = {}
+    t1: Dict[int, float] = {}
+    t2: Dict[int, float] = {}
+    move_fidelity: Dict[Tuple[int, int], float] = {}
+    quality_metric_set = IQMClient(iqm_server_url, quantum_computer=quantum_computer).get_quality_metric_set()
+    calibration_metrics = quality_metric_set.observations
+
     list_couplings = []
     list_fids = []
-    if "double_move_gate_fidelity" in cal_keys.keys():
-        i, j = cal_keys["double_move_gate_fidelity"]
+    gates_info: Dict[str, Dict[str, Any]] = {}
+    for gate in backend.architecture.gates.keys():
+        gates_info[gate] = {
+            x: backend.architecture.gates[gate].implementations[x].loci
+            for x in backend.architecture.gates[gate].implementations.keys()
+        }
+
+    if backend.has_resonators():
         topology = "star"
     else:
-        i, j = cal_keys["cz_gate_fidelity"]
         topology = "crystal"
-    for item in calibration["calibrations"][i]["metrics"][j]["metrics"]:
-        qb1 = (
-            int(item["locus"][0][2:])
-            if not any(resonator in item["locus"][0] for resonator in resonator_names)
-            else 0
-        )
-        qb2 = (
-            int(item["locus"][1][2:])
-            if not any(resonator in item["locus"][1] for resonator in resonator_names)
-            else 0
-        )
-        list_couplings.append([qb1, qb2])
-        list_fids.append(float(item["value"]))
-    calibrated_qubits = set(np.array(list_couplings).reshape(-1))
 
-    # Process all metrics if all_metrics is True
-    metrics_dict: Dict[str, Dict[Union[int, Tuple[int, int]], float]] = {}
-    for metric_key, (i, j) in cal_keys.items():
-        metric_data = calibration["calibrations"][i]["metrics"][j]["metrics"]
-        metrics_dict[metric_key] = {}
+        # Iterate over the calibration metrics
+    for metrics in calibration_metrics:
+        dut_field = metrics.dut_field
+        values = metrics.value
+        if all(obs in dut_field for obs in ObservationType.READOUT.value):
+            qubit_index = int(dut_field.split("QB")[1].split(".")[0])
+            readout_fidelity[qubit_index] = values
+        elif ObservationType.SQG.value in dut_field and any(
+            x in dut_field for x in gates_info[ObservationType.SQG.value]
+        ):
+            qubit_index = int(dut_field.split("QB")[1].split(".")[0])
+            single_qubit_fidelity[qubit_index] = values
+        elif all(obs in dut_field for obs in ObservationType.T1.value):
+            qubit_index = int(dut_field.split("QB")[1].split(".")[0])
+            t1[qubit_index] = values * 10**6
+        elif all(obs in dut_field for obs in ObservationType.T2.value):
+            qubit_index = int(dut_field.split("QB")[1].split(".")[0])
+            t2[qubit_index] = values * 10**6
+        elif ObservationType.DOUBLE_MOVE.value in dut_field and any(
+            x in dut_field for x in gates_info[ObservationType.DOUBLE_MOVE.value]
+        ):
+            qb_matches = re.findall(r"QB\d+", dut_field)
+            qbx = int(qb_matches[0].split("QB")[1])
+            move_fidelity[(qbx, 0)] = values
+            move_fidelity[(0, qbx)] = values
+        elif (
+            ObservationType.CZ.value in dut_field
+            and any(x in dut_field for x in gates_info[ObservationType.CZ.value])
+            and backend.has_resonators()
+        ):
+            qb_matches = re.findall(r"QB\d+", dut_field)
+            qbx = int(qb_matches[0].split("QB")[1])
+            list_couplings.append([qbx, 0])
+            list_fids.append(values)
+            cz_fidelity[(qbx, 0)] = values
+            cz_fidelity[(0, qbx)] = values
+        elif ObservationType.CZ.value in dut_field and any(
+            x in dut_field for x in gates_info[ObservationType.CZ.value]
+        ):
+            qb_matches = re.findall(r"QB\d+", dut_field)
+            qbx, qby = int(qb_matches[0].split("QB")[1]), int(qb_matches[1].split("QB")[1])
+            cz_fidelity[(qbx, qby)] = values
+            cz_fidelity[(qby, qbx)] = values
+            list_couplings.append([qbx, qby])
+            list_fids.append(values)
 
-        for item in metric_data:
-            # Determine if it's a single or two-qubit metric
-            if "component" in item:
-                # Single qubit metric
-                component = item["component"]
-                if not any(resonator in component for resonator in resonator_names):
-                    qb = int(component[2:])
-                    metrics_dict[metric_key][qb] = float(item["value"])
-                    calibrated_qubits.add(qb) # Add qubits that have a single qubit metric
-            if "locus" in item and len(item["locus"]) == 2:
-                # Two qubit metric
-                locus = item["locus"]
-                if not (any(resonator in locus[0] for resonator in resonator_names) or any(resonator in locus[1] for resonator in resonator_names)):
-                    qb1 = int(locus[0][2:])
-                    qb2 = int(locus[1][2:])
-                    metrics_dict[metric_key][(qb1, qb2)] = float(item["value"])
-
+    metrics_dict: Dict[str, Dict[Union[int, Tuple[int, int]], float]] = {
+        "cz_gate_fidelity": cz_fidelity,
+        "fidelity_1qb_gates_averaged": single_qubit_fidelity,
+        "single_shot_readout_fidelity": readout_fidelity,
+        "t1_time": t1,
+        "t2_time": t2,
+        "double_move_gate_fidelity": move_fidelity,
+    }
     # Enumerate all calibrated qubits starting from 0
+    calibrated_qubits = set(np.array(list_couplings).reshape(-1))
     qubit_mapping = {qubit: idx for idx, qubit in enumerate(calibrated_qubits)}
     list_couplings = [[qubit_mapping[edge[0]], qubit_mapping[edge[1]]] for edge in list_couplings]
-
     # Apply the qubit mapping to metrics_dict
     remapped_metrics_dict = {}
     for metric_key, metric_values in metrics_dict.items():
@@ -383,10 +418,6 @@ def extract_fidelities(cal_url: str, all_metrics: bool = False) -> Union[
                 # Single-qubit metric
                 remapped_metrics_dict[metric_key][qubit_mapping[key]] = value
     metrics_dict = remapped_metrics_dict
-
-    # If all_metrics is False, only return everything related to CZ fidelites
-    if not all_metrics:
-        return list_couplings, list_fids, topology, qubit_mapping
 
     return list_couplings, list_fids, topology, qubit_mapping, metrics_dict
 
@@ -606,6 +637,7 @@ def perform_backend_transpilation(
     optimize_sqg: bool = False,
     drop_final_rz: bool = True,
     routing_method: Optional[str] = "sabre",
+    approximation_degree: float = 1.0,
 ) -> List[QuantumCircuit]:
     """
     Transpile a list of circuits to backend specifications.
@@ -620,6 +652,7 @@ def perform_backend_transpilation(
         optimize_sqg (bool): Whether SQG optimization is performed taking into account virtual Z.
         drop_final_rz (bool): Whether the SQG optimizer drops a final RZ gate.
         routing_method (Optional[str]): The routing method employed by Qiskit's transpilation pass.
+        approximation_degree (int): The target fidelity to which arbitrary two qubit gates are approximated.
 
     Returns:
         List[QuantumCircuit]: A list of transpiled quantum circuits.
@@ -641,6 +674,7 @@ def perform_backend_transpilation(
                 optimize_single_qubits=optimize_sqg,
                 remove_final_rzs=drop_final_rz,
                 coupling_map=coupling_map_red,
+                approximation_degree=approximation_degree,
                 # initial_layout=qubits if aux_qc is None else None,
             )
         else:
@@ -651,6 +685,7 @@ def perform_backend_transpilation(
                 optimization_level=qiskit_optim_level,
                 initial_layout=qubits if aux_qc is None else None,
                 routing_method=routing_method,
+                approximation_degree=approximation_degree,
             )
             if aux_qc is not None:
                 transpiled = aux_qc.compose(transpiled, qubits=qubits, clbits=list(range(qc.num_clbits)))
@@ -797,31 +832,45 @@ def retrieve_all_counts(iqm_jobs: List[IQMJob], identifier: Optional[str] = None
 
 
 def retrieve_all_job_metadata(
-    iqm_jobs: List[IQMJob],
-) -> Dict[str, Dict[str, Any]]:
-    """Retrieve the counts from a list of IQMJob objects.
+    iqm_jobs: list[IQMJob],
+) -> dict[str, dict[str, Any]]:
+    """Retrieve the metadata from a list of Job objects.
+
     Args:
-        iqm_jobs List[IQMJob]: The list of IQMJob objects.
+        iqm_jobs : List[IQMJob]
+            The list of IQMJob objects.
 
     Returns:
         Dict[str, Dict[str, Any]]: Relevant metadata of all the IQMJob objects.
+
     """
     all_meta = {}
-
     for index, j in enumerate(iqm_jobs):
         all_attributes_j = dir(j)
+        if not hasattr(j, "_iqm_job"):
+            shots = j.metadata["shots"] if "shots" in j.metadata.keys() else None
+            timestamps = j.metadata["timestamps"] if "timestamps" in j.metadata.keys() else None
+        else:
+            job_parameters = j._iqm_job._parameters
+            if job_parameters is not None:
+                shots = job_parameters.shots if "shots" in job_parameters.__dict__.keys() else None
+            else:
+                raise ValueError("Job parameters return None, cannot retrieve shots information.")
+            timestamps = {}
+            for entry in j._iqm_job.data.timeline:
+                timestamps.update({entry.status: entry.timestamp})
         all_meta.update(
             {
                 "batch_job_"
                 + str(index + 1): {
                     "job_id": j.job_id() if "job_id" in all_attributes_j else None,
-                    "backend": j.backend().name if "backend" in all_attributes_j else None,
-                    "status": j.status().value if "status" in all_attributes_j else None,
+                    "backend": (j.backend().name if "backend" in all_attributes_j else None),
+                    "status": (j.status().value if "status" in all_attributes_j else None),
                     "circuits_in_batch": (
-                        len(cast(List, j.circuit_metadata)) if "circuit_metadata" in all_attributes_j else None
+                        len(cast(list, j.circuit_metadata)) if "circuit_metadata" in all_attributes_j else None
                     ),
-                    "shots": j.metadata["shots"] if "shots" in j.metadata.keys() else None,
-                    "timestamps": j.metadata["timestamps"] if "timestamps" in j.metadata.keys() else None,
+                    "shots": shots,
+                    "timestamps": timestamps,
                 }
             }
         )