+++ CHANGELOG.rst
@@ -2,6 +2,12 @@
 Changelog
 =========
 
+Version 2.55
+============
+* Compatibility with new job metadata formats.
+* Unified fetching of calibration data via IQMClient.
+* Addition of active reset and approximate transpilation options in the configurations.
+
 Version 2.54
 ============
 * Update iqm-client and iqm-station-control-client dependency ranges for compatibility with IQM OS 4.4.